Add files via upload
This file contains the notebook on which the VGG-16 CNN Model was trained on for our leaf dataset to detect the plant health condition. Our model classifies the plant leaf into 4 major categories which include the following [Healthy, Multiple Diseases, Scab, Rust] . The model requires a powerful GPU and hence was run using google colab pro resources. The model achieved a 80% training accuracy and a 70% validation accuracy.